test(core): split work authority surface responsibilities
(cherry picked from commit ad199dbd3b0272ed78340c09ec20028474c2e63a)
