+++ .buildchain/kfd/kfd-2/release-claims.json
@@ -11,7 +11,7 @@
     "version": "4.0.0-alpha.5",
     "channel": "repository-default",
     "tag": "v4.0.0-alpha.5",
-    "sourceSha": "8c3588ff5376ba63ae769d42008f5168c49ec568"
+    "sourceSha": "91cb0f8074bba2307a444451a4550cc538928531"
   },
   "claims": [
     {
