+++ framework/core/tests/python/_work_authority_surface_review_cases.py
@@ -1,0 +1,427 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Independent review and close-planning cases."""
+# ruff: noqa: F401,F403
+
+from _work_authority_surface_support import *
+
+
+def test_reviewer_result_requires_exact_criterion_coverage():
+    checks = ["Names the product", "Uses retained evidence"]
+    report = {
+        "providerObservation": {
+            "text": (
+                "Review complete.\n"
+                "KUNGFU_REVIEW_RESULT "
+                '{"verdict":"fit","summary":"Both checks pass.",'
+                '"criteria":['
+                '{"criterion":"Names the product","passed":true,'
+                '"evidence":"deliverable line 1"},'
+                '{"criterion":"Uses retained evidence","passed":true,'
+                '"evidence":"inputs and benefit section"}],'
+                '"evidenceRequests":[]}'
+            )
+        }
+    }
+    result = assignment_review.parse_reviewer_result(report, checks)
+    assert result["verdict"] == "fit"
+    assert all(row["passed"] for row in result["criteria"])
+
+
+def test_reviewer_prompt_requires_exact_structured_criterion_coverage():
+    checks = ["Names the product", "Reports evidence and unresolved risks"]
+    profile_root = f"sha256:{'1' * 64}"
+    assignment_root = f"sha256:{'2' * 64}"
+    query_proof_root = f"sha256:{'3' * 64}"
+    report_root = f"sha256:{'5' * 64}"
+    deliverable_root = f"sha256:{'4' * 64}"
+    plan = {
+        "workspace": {"id": "project:starter"},
+        "work": {
+            "initiativeId": "starter",
+            "assignmentId": "write-brief",
+            "assignmentRoot": assignment_root,
+            "queryProofRoot": query_proof_root,
+            "acceptanceChecks": checks,
+        },
+        "execution": {
+            "reportRoot": report_root,
+            "workRef": {"profileRoot": profile_root},
+        },
+        "deliverable": {"path": "README.md", "root": deliverable_root},
+        "inputs": [],
+    }
+    prompt = assignment_review.review_agent_prompt(plan)
+    continuation = assignment_review.review_continuation(plan)
+
+    assert "criteria array must contain exactly 2 objects" in prompt
+    assert '1. criterion must equal "Names the product"' in prompt
+    assert '2. criterion must equal "Reports evidence and unresolved risks"' in prompt
+    assert "A statement in summary does not count as criterion coverage" in prompt
+    assert '"schema": "kungfu.review-context/v1"' in prompt
+    assert '"schema": "kungfu.work-ref/v1"' in prompt
+    assert '"entityId": "write-brief"' in prompt
+    assert '"initiativeId": "starter"' in prompt
+    assert f'"profileRoot": "{profile_root}"' in prompt
+    assert f'"systemTimeCut": "{query_proof_root}"' in prompt
+    assert '"mode": "fresh-independent-review"' in prompt
+    assert f'"priorClaimRoot": "{report_root}"' in prompt
+    assert '"schema": "kungfu.agent-continuation-envelope/v1"' in prompt
+    assert '"schema": "kungfu.review-intake-assessment/v1"' in prompt
+    assert "assess that admitted content directly" in prompt
+    assert "only for evidence rows that do not contain content" in prompt
+    assert run_agent.validate_continuation(continuation) == continuation
+    assert session_contract.validate_work_ref(
+        assignment_review.review_work_ref(plan)
+    ) == assignment_review.review_work_ref(plan)
+
+
+def test_fresh_reviewer_receives_the_exact_work_continuation_envelope(
+    tmp_path, monkeypatch
+):
+    def root(marker):
+        return f"sha256:{marker * 64}"
+
+    plan = {
+        "planRoot": root("9"),
+        "workspace": {"id": "project:starter"},
+        "work": {
+            "initiativeId": "starter",
+            "assignmentId": "write-brief",
+            "assignmentRoot": root("1"),
+            "queryProofRoot": root("2"),
+            "acceptanceChecks": ["Names the product"],
+            "phase": "executing",
+        },
+        "execution": {
+            "reportRoot": root("3"),
+            "workRef": {"profileRoot": root("4")},
+        },
+        "deliverable": {"path": "README.md", "root": root("5")},
+        "inputs": [],
+        "reviewer": {"id": "codex.path", "label": "Codex · PATH CLI"},
+        "reviewExecution": {"mode": "fresh-process"},
+        "executable": True,
+    }
+    observed = {}
+
+    def execute_agent(**kwargs):
+        observed.update(kwargs)
+        return {
+            "launch": {"exitCode": 1},
+            "reportRoot": root("6"),
+        }
+
+    monkeypatch.setattr(assignment_review_lifecycle.run_agent, "execute", execute_agent)
+    services = assignment_review_lifecycle.ReviewServices(
+        plan=lambda **_kwargs: plan,
+        receipt=lambda value: value,
+        runtime=lambda *_args: SimpleNamespace(
+            identity=SimpleNamespace(workspace_root=str(tmp_path)),
+            runtime_dir=str(tmp_path / "runtime"),
+        ),
+        retained_evidence=lambda *_args: None,
+        agent_report_summary=lambda report: report,
+        status=lambda *_args: {},
+        mint_lease=lambda *_args: {},
+        advance=lambda *_args: {},
+        completion_claim_values=lambda *_args: {},
+        profile_action=lambda *_args: {},
+        completion_review_values=lambda *_args: {},
+    )
+    request = assignment_review_lifecycle.ReviewRequest(
+        config_home=tmp_path / "config",
+        runtime_home=tmp_path / "home",
+        agent_report_file=tmp_path / "execution-report.json",
+        workspace_root=str(tmp_path),
+        home=False,
+        initiative_id="starter",
+        assignment_id="write-brief",
+        reviewer_profile_id="codex.path",
+        expected_plan_root=plan["planRoot"],
+        execute=True,
+        allow_foreign_binding=False,
+    )
+
+    receipt = assignment_review_lifecycle.execute(
+        request, services, lambda *_args: None, lambda: 0
+    )
+
+    expected_work_ref = assignment_review.review_work_ref(plan)
+    expected_continuation = assignment_review.review_continuation(plan)
+    assert receipt["status"] == "reviewer-failed"
+    assert observed["work_ref"] == expected_work_ref
+    assert observed["continuation"] == expected_continuation
+    assert observed["permission_mode"] == "read-only"
+    assert (
+        run_agent.validate_continuation(observed["continuation"])
+        == expected_continuation
+    )
+    assert json.dumps(expected_continuation, sort_keys=True) in observed["prompt"]
+
+
+def test_exact_retained_passing_reviewer_evidence_is_reusable(tmp_path):
+    command = importlib.import_module("kungfu.cli.commands.assignment")
+    runtime = tmp_path / "runtime"
+    workspace = tmp_path / "project"
+    report_path = runtime / "agent-runs" / "review-run" / "bundle" / "report.json"
+    report_path.parent.mkdir(parents=True)
+    workspace.mkdir()
+    checks = ["Names the product", "Uses retained evidence"]
+    plan = {
+        "workspace": {"id": "project:starter", "root": str(workspace)},
+        "work": {
+            "initiativeId": "starter",
+            "assignmentId": "write-brief",
+            "assignmentRoot": f"sha256:{'1' * 64}",
+            "queryProofRoot": f"sha256:{'7' * 64}",
+            "acceptanceChecks": checks,
+        },
+        "deliverable": {
+            "path": "deliverables/launch-brief.md",
+            "root": f"sha256:{'2' * 64}",
+        },
+        "inputs": [
+            {
+                "path": "inputs/product-notes.md",
+                "root": f"sha256:{'3' * 64}",
+            }
+        ],
+        "execution": {
+            "reportRoot": f"sha256:{'8' * 64}",
+            "workRef": {"profileRoot": f"sha256:{'4' * 64}"},
+        },
+        "reviewer": {
+            "id": "codex.local",
+            "profileRoot": f"sha256:{'5' * 64}",
+        },
+    }
+    result_line = {
+        "verdict": "fit",
+        "summary": "Both checks pass.",
+        "criteria": [
+            {
+                "criterion": "Names the product",
+                "passed": True,
+                "evidence": "deliverable line 1",
+            },
+            {
+                "criterion": "Uses retained evidence",
+                "passed": True,
+                "evidence": "input line 2",
+            },
+        ],
+        "evidenceRequests": [],
+    }
+    report_body = {
+        "schema": run_agent.REPORT_SCHEMA,
+        "runId": "review-run",
+        "attemptId": "review-run",
+        "runtimeProfile": {
+            "id": "codex.local",
+            "root": plan["reviewer"]["profileRoot"],
+        },
+        "launch": {
+            "cwd": str(workspace),
+            "argvWithoutPrompt": ["codex", "exec", "--sandbox", "read-only"],
+            "promptRoot": run_agent.canonical_root(
+                assignment_review.review_agent_prompt(plan)
+            ),
+            "exitCode": 0,
+        },
+        "providerObservation": {
+            "text": "KUNGFU_REVIEW_RESULT "
+            + json.dumps(result_line, separators=(",", ":"))
+        },
+        "work": {
+            "workRef": {
+                "workspaceId": plan["workspace"]["id"],
+                "profileId": "kungfu.work-control",
+                "profileRoot": plan["execution"]["workRef"]["profileRoot"],
+                "entityType": "assignment",
+                "entityId": plan["work"]["assignmentId"],
+                "entityRoot": plan["work"]["assignmentRoot"],
+                "purpose": "independent-completion-review",
+                "systemTimeCut": f"sha256:{'6' * 64}",
+                "initiativeId": plan["work"]["initiativeId"],
+            }
+        },
+        "privacy": {
+            "priorTranscriptBytesGivenToAgent": 0,
+            "privateProviderSessionStoreRead": False,
+        },
+        "episode": {"episodeId": "42"},
+    }
+    report = {
+        **report_body,
+        "reportRoot": run_agent.canonical_root(report_body),
+    }
+    report_path.write_text(json.dumps(report), encoding="utf-8")
+
+    retained = command._find_retained_reviewer_evidence(runtime, plan)
+
+    assert retained is not None
+    assert retained["report"]["reportRoot"] == report["reportRoot"]
+    assert retained["assessment"]["verdict"] == "fit"
+
+    changed_plan = {
+        **plan,
+        "deliverable": {
+            **plan["deliverable"],
+            "root": f"sha256:{'7' * 64}",
+        },
+    }
+    assert command._find_retained_reviewer_evidence(runtime, changed_plan) is None
+
+
+def test_review_settlement_lease_renews_the_original_execution_agent(monkeypatch):
+    command = importlib.import_module("kungfu.cli.commands.assignment")
+    observed = {}
+
+    def profile_action(runtime_dir, operation, values, actor):
+        observed.update(
+            {
+                "runtime_dir": runtime_dir,
+                "operation": operation,
+                "values": values,
+                "actor": actor,
+            }
+        )
+        return {"claim": values}
+
+    monkeypatch.setattr(command, "_profile_action", profile_action)
+    before = datetime.now(UTC)
+    receipt = command._mint_review_settlement_lease(
+        "/runtime",
+        {
+            "work": {
+                "initiativeId": "starter",
+                "assignmentId": "write-brief",
+            },
+            "execution": {"agent": {"id": "codex.execution", "provider": "codex"}},
+        },
+        "local-user",
+    )
+    expiry = datetime.fromisoformat(
+        observed["values"]["leaseExpiresAt"].replace("Z", "+00:00")
+    )
+
+    assert receipt["claim"]["agent"] == "codex.execution"
+    assert observed["operation"] == "claim-assignment"
+    assert observed["values"]["slot"] == "starter-codex"
+    assert observed["values"]["grantScope"] == "assignment-execution"
+    assert observed["values"]["leaseId"].startswith("work-review-")
+    assert expiry > before
+
+
+def test_starter_native_review_uses_the_bound_handoff_assessment_without_git():
+    command = importlib.import_module("kungfu.cli.commands.assignment")
+    values = command._native_completion_review_values(
+        {
+            "workspace": {"root": "/project-without-git"},
+            "work": {
+                "initiativeId": "starter",
+                "assignmentId": "write-brief",
+            },
+            "reviewer": {"id": "codex.path.detected"},
+        },
+        {"runId": "review-run"},
+    )
+
+    assert values["purpose"] == "handoff"
+    assert values["executorProfile"] == "thread"
+    assert values["reviewer"] == "codex.path.detected"
+    assert values["reviewerSource"] == "review-run"
+    assert values["checkoutPath"] == ""
+
+
+def test_starter_completion_claim_uses_native_episode_and_proof_authority():
+    command = importlib.import_module("kungfu.cli.commands.assignment")
+    values = command._native_completion_claim_values(
+        {
+            "work": {
+                "initiativeId": "starter",
+                "assignmentId": "write-brief",
+                "workDefinitionRoot": f"sha256:{'1' * 64}",
+                "acceptanceChecks": ["Names the product"],
+            },
+            "execution": {
+                "episodeId": "41",
+                "reportRoot": f"sha256:{'2' * 64}",
+            },
+            "deliverable": {"root": f"sha256:{'3' * 64}"},
+        },
+        {
+            "runId": "review-run",
+            "reportRoot": f"sha256:{'4' * 64}",
+            "episode": {"episodeId": "42"},
+        },
+        {"summary": "The acceptance criterion passes."},
+    )
+
+    assert values["evidenceEpisodeIds"] == [41, 42]
+    assert values["proofRoots"] == [
+        f"sha256:{'2' * 64}",
+        f"sha256:{'4' * 64}",
+        f"sha256:{'3' * 64}",
+    ]
+    assert "resultAtlasRoot" not in values
+    assert "inputAtlasRoot" not in values
+    assert "projectCutRoot" not in values
+
+
+def test_starter_close_plan_binds_one_fit_review_and_hides_technical_seal(
+    monkeypatch,
+):
+    command = importlib.import_module("kungfu.cli.commands.assignment")
+    review = {
+        "review_id": "review-starter",
+        "verdict": "fit",
+        "continuation_plan": {
+            "allowed_actions": ["approve", "close"],
+        },
+        "continuation_plan_root": f"sha256:{'1' * 64}",
+    }
+    identity = type(
+        "Identity",
+        (),
+        {
+            "workspace_id": "project:starter",
+            "workspace_root": "/project",
+            "data_home": "/project/.kungfu",
+            "identity_root": f"sha256:{'2' * 64}",
+        },
+    )()
+    monkeypatch.setattr(
+        command,
+        "_runtime",
+        lambda *args: (identity, "/project/.kungfu/runtime", {}),
+    )
+    monkeypatch.setattr(
+        command,
+        "_status",
+        lambda *args: {
+            "phase": "independently-reviewed",
+            "query_proof_root": f"sha256:{'3' * 64}",
+            "assignment": {"assignment_id": "write-brief"},
+            "independent_reviews": [review],
+            "continuation_decisions": [],
+        },
+    )
+
+    plan = assignment_close.build_plan(
+        workspace_root="/project",
+        home=False,
+        initiative_id="starter",
+        assignment_id="write-brief",
+        services=command._close_services(),
+    )
+
+    assert plan["writeOccurred"] is False
+    assert plan["confirmationRequired"] is True
+    assert plan["executable"] is True
+    assert plan["decision"]["mode"] == "required"
+    assert [effect["stage"] for effect in plan["effects"]] == [
+        "decide",
+        "seal",
+    ]
+    assert plan["review"]["allowedActions"] == ["approve", "close"]
